+++ src/windows/wsl/CMakeLists.txt
@@ -9,8 +9,11 @@ add_executable(wsl ${SOURCES} ${HEADERS})
 
 target_link_libraries(wsl
                       ${COMMON_LINK_LIBRARIES}
+                      ${MSI_LINK_LIBRARIES}
                       common
-                      runtimeobject.lib)
+                      runtimeobject.lib
+                      delayimp.lib)
 
+set_target_properties(wsl PROPERTIES LINK_FLAGS "/DELAYLOAD:msi.dll /DELAYLOAD:WINTRUST.dll")
 target_precompile_headers(wsl REUSE_FROM common)
 set_target_properties(wsl PROPERTIES FOLDER windows)
