+++ test/windows/CMakeLists.txt
@@ -24,6 +24,9 @@ target_link_libraries(wsltests
                       common
                       ${TAEF_LINK_LIBRARIES}
                       ${COMMON_LINK_LIBRARIES}
+                      ${MSI_LINK_LIBRARIES}
+                      ${HCS_LINK_LIBRARIES}
+                      ${SERVICE_LINK_LIBRARIES}
                       VirtDisk.lib
                       Wer.lib
                       Dbghelp.lib
